Crate faerie

Faerie is a crate for creating object files.

Re-exports

pub use crate::artifact::Artifact;

pub use crate::artifact::ArtifactBuilder;

pub use crate::artifact::ArtifactError;

pub use crate::artifact::Data;

pub use crate::artifact::Link;

pub use crate::artifact::Reloc;

Modules

artifact
An artifact is a platform independent binary object file format abstraction.

Structs

DataDecl
Builder for data declarations

DataImportDecl
Builder for data import declarations

FunctionDecl
Builder for function declarations

FunctionImportDecl
Builder for function import declarations

SectionDecl
Builder for a section declaration

Enums

DataType
Type of data declared

Decl
The kind of declaration this is

ImportKind
The kind of import this is - either a function, or a copy relocation of data from a shared library

Scope
Linker binding scope of a definition

SectionKind
The kind of this section

Visibility
Linker visibility of a definition