Module faerie::artifact

An artifact is a platform independent binary object file format abstraction.

Structs

Artifact
An abstract binary artifact, which contains code, data, imports, and relocations

ArtifactBuilder
Builder for creating an artifact

Binding
A binding of a raw name to its declaration, decl

Link
An abstract relocation linking one symbol to another, at an offset

LinkAndDecl
A relocation binding one declaration to another

Enums

ArtifactError
The kinds of errors that can befall someone creating an Artifact

Data
The data to be stored in an artifact, representing a function body or data object.

DataType
Type of data declared

Decl
The kind of declaration this is

DefinedDecl
A declaration that is defined inside this artifact

ImportKind
The kind of import this is - either a function, or a copy relocation of data from a shared library

Reloc
The kind of relocation for a link.

Scope
Linker binding scope of a definition

SectionKind
The kind of this section

Visibility
Linker visibility of a definition