Enum faerie::artifact::Decl

pub enum Decl {
    Import(ImportKind),
    Defined(DefinedDecl),
}

The kind of declaration this is

Variants

Import(ImportKind)

Declaration of an import

Defined(DefinedDecl)

Declaration of an item to be defined in this artifact

Methods

impl Decl

pub fn function_import() -> FunctionImportDecl

An import of a function/routine defined in a shared library

pub fn data_import() -> DataImportDecl

A GOT-based import of data defined in a shared library

pub fn function() -> FunctionDecl

A function defined in this artifact

pub fn data() -> DataDecl

A data object defined in this artifact

pub fn cstring() -> DataDecl

A null-terminated string object defined in this artifact

pub fn section(kind: SectionKind) -> SectionDecl

A section defined in this artifact

pub fn absorb(&mut self, other: Self) -> Result<(), ArtifactError>

If it is compatible, absorb the new declaration (other) into the old (self); otherwise returns an error.

The rule here is "C-ish", but essentially:

Duplicate declarations are no-ops / ignored.

If the previous declaration was an FunctionImport or DataImport, then if the subsequent declaration is a corresponding matching Function or Data declaration, it is said to be "upgraded", and forever after is considered a declaration in need of a definition.

If the previous declaration was a Function or Data declaration, then a subsequent corresponding FunctionImport or DataImport is a no-op.

Anything else is a IncompatibleDeclaration error!

pub fn is_import(&self) -> bool

Is this an import (function or data) from a shared library?

pub fn is_section(&self) -> bool

Is this a section?